envs/devpulse_code_review_env/server/requirements.txt
agno
openai
httpx
feedparser
streamlit>=1.30
pydantic
openenv-core